Rename the typing module
