+++ docs/README.md
@@ -41,16 +41,22 @@ In the realm of the workflow management systems, there are well defined inputs a
 `semantikon` provides a way to define types for any number of input parameters and any number of output values for function via type hinting, in particular: data type, unit and ontological type. Type hinting is done with the function `u`, which **requires** the type, and **optionally** you can define the units and the ontological type. The type hinting is done in the following way:
 
 ```python
->>> from semantikon.typing import u
->>> from rdflib import Namespace
->>>
->>> EX = Namespace("http://example.org/")
->>> 
->>> def get_speed(
-...     distance: u(float, units="meter", uri=EX.distance),
-...     time: u(float, units="second", uri=EX.time),
+>> > from semantikon.metadata import u
+>> > from rdflib import Namespace
+>> >
+>> > EX = Namespace("http://example.org/")
+>> >
+>> >
+
+def get_speed(
+        ...distance: u(float, units="meter", uri=EX.distance),
+
+
+...
+time: u(float, units="second", uri=EX.time),
 ... ) -> u(float, units="meter/second", label="speed", uri=EX.speed):
-...     return distance / time
+...
+return distance / time
 
 ```
 
@@ -58,22 +64,38 @@ In the realm of the workflow management systems, there are well defined inputs a
 
 You can also type-hint the inputs and outputs of a function using a class, i.e.:
 
-
 ```python
->>> from semantikon.typing import u
->>> from semantikon.converter import semantikon_class
->>> from rdflib import Namespace
->>>
->>> EX = Namespace("http://example.org/")
->>> 
->>> @semantikon_class
-... class MyRecord:
-...     distance: u(float, units="meter", uri=EX.distance)
-...     time: u(float, units="second", uri=EX.time)
-...     result: u(float, units="meter/second", label="speed", uri=EX.speed)
->>> 
->>> def get_speed(distance: MyRecord.distance, time: MyRecord.time) -> MyRecord.result:
-...     return distance / time
+>> > from semantikon.metadata import u
+>> > from semantikon.converter import semantikon_class
+>> > from rdflib import Namespace
+>> >
+>> > EX = Namespace("http://example.org/")
+>> >
+>> >
+
+@semantikon_class
+
+
+...
+
+
+class MyRecord:
+
+
+    ...
+distance: u(float, units="meter", uri=EX.distance)
+...
+time: u(float, units="second", uri=EX.time)
+...
+result: u(float, units="meter/second", label="speed", uri=EX.speed)
+>> >
+>> >
+
+def get_speed(distance: MyRecord.distance, time: MyRecord.time) -> MyRecord.result:
+
+
+    ...
+return distance / time
 
 ```
 
@@ -90,23 +112,39 @@ In order to extract argument information, you can use the functions `parse_input
 Example:
 
 ```python
->>> from semantikon.typing import u
->>> from semantikon.converter import parse_input_args, parse_output_args
->>> from rdflib import Namespace
->>>
->>> EX = Namespace("http://example.org/")
->>> 
->>> def get_speed(
-...     a: u(float, units="meter", uri=EX.distance),
-...     b: u(float, units="second", uri=EX.time),
-... ) -> u(float, units="meter/second", label="speed", uri=EX.speed):
-...     return a / b
->>> 
->>> print(parse_input_args(get_speed))
-{'a': {'units': 'meter', 'uri': rdflib.term.URIRef('http://example.org/distance'), 'dtype': <class 'float'>}, 'b': {'units': 'second', 'uri': rdflib.term.URIRef('http://example.org/time'), 'dtype': <class 'float'>}}
+>> > from semantikon.metadata import u
+>> > from semantikon.converter import parse_input_args, parse_output_args
+>> > from rdflib import Namespace
+>> >
+>> > EX = Namespace("http://example.org/")
+>> >
+>> >
 
->>> print(parse_output_args(get_speed))
-{'units': 'meter/second', 'label': 'speed', 'uri': rdflib.term.URIRef('http://example.org/speed'), 'dtype': <class 'float'>}
+def get_speed(
+        ...a: u(float, units="meter", uri=EX.distance),
+
+
+...
+b: u(float, units="second", uri=EX.time),
+... ) -> u(float, units="meter/second", label="speed", uri=EX.speed):
+...
+return a / b
+>> >
+>> > print(parse_input_args(get_speed))
+{'a': {'units': 'meter', 'uri': rdflib.term.URIRef('http://example.org/distance'), 'dtype': <
+
+
+class 'float'>}, 'b': {'units': 'second', 'uri': rdflib.term.URIRef('http://example.org/time'), 'dtype': <
+
+
+class 'float'>}}
+
+>> > print(parse_output_args(get_speed))
+{'units': 'meter/second', 'label': 'speed'
+
+, 'uri': rdflib.term.URIRef('http://example.org/speed'), 'dtype': <
+
+class 'float'>}
 
 ```
 
@@ -115,21 +153,33 @@ Example:
 `semantikon` provides a way to interpret the types of inputs and outputs of a function via a decorator, in order to check consistency of the types and to convert them if necessary. Currently, `semantikon` provides an interpreter for `pint.UnitRegistry` objects. The interpreter is applied in the following way:
 
 ```python
->>> from semantikon.typing import u
->>> from semantikon.converter import units
->>> from pint import UnitRegistry
->>> 
->>> @units
-... def get_speed(
-...     a: u(float, units="meter"),
-...     b: u(float, units="second")
+>> > from semantikon.metadata import u
+>> > from semantikon.converter import units
+>> > from pint import UnitRegistry
+>> >
+>> >
+
+@units
+
+
+...
+
+
+def get_speed(
+        ...a: u(float, units="meter"),
+
+
+...
+b: u(float, units="second")
 ... ) -> u(float, units="meter/second", label="speed"):
-...     return a / b
->>> 
->>> ureg = UnitRegistry()
->>> 
->>> print(get_speed(1 * ureg.meter, 1 * ureg.second))
-1.0 meter / second
+...
+return a / b
+>> >
+>> > ureg = UnitRegistry()
+>> >
+>> > print(get_speed(1 * ureg.meter, 1 * ureg.second))
+1.0
+meter / second
 
 ```
