+++ tests/unit/test_units.py
@@ -3,7 +3,7 @@ import unittest
 from pint import UnitRegistry
 
 from semantikon.converter import units
-from semantikon.typing import u
+from semantikon.metadata import u
 
 
 @units
